DataAnalytics/VideoAnalytics/va_requirements.txt
numpy==1.14.5
opencv-python==3.4.1.15
pyzmq==18.0.2
